requirements.txt
fastapi
python-multipart
PyMuPDF
openai
python-dotenv
tenacity
pandas
openpyxl
jinja2
